qsl-auto-v2/requirements.txt
pydantic>=2.7
pydantic-settings>=2.2
jinja2>=3.1
weasyprint==62.3
html5lib>=1.1
cairosvg>=2.7
cairocffi>=1.6
tinycss2>=1.2
cssselect2>=0.7

httpx>=0.27
fastapi>=0.111
uvicorn[standard]>=0.30

google-api-python-client>=2.119.0
google-auth>=2.31.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

Pillow>=10.3
PyPDF2>=3.0

Typer>=0.12
rich>=13.7
python-dotenv>=1.0
structlog>=24.1

pytest>=8.2
pytest-asyncio>=0.23
